Move interfaces into own files
