+++ dependency-reduced-pom.xml
@@ -2,10 +2,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>hwr.oop.projects</groupId>
-  <artifactId>project</artifactId>
-  <name>project</name>
-  <version>0.1.0-SNAPSHOT</version>
-  <description>project-description</description>
+  <artifactId>com.group1.allinorfold</artifactId>
+  <name>AllInOrFold</name>
+  <version>0.0.1</version>
+  <description>A poker cli game</description>
   <build>
     <sourceDirectory>src/main/kotlin</sourceDirectory>
     <testSourceDirectory>src/test/kotlin</testSourceDirectory>
