+++ src/main/kotlin/hwr/oop/group1/poker/cli/StateManager.kt
@@ -1,0 +1,6 @@
+package hwr.oop.group1.poker.cli
+
+interface StateManager {
+    fun saveState(state: Map<String, Any>)
+    fun loadState(): Map<String, Any>?
+}
