+++ src/main/kotlin/hwr/oop/group1/poker/cli/StateSerializable.kt
@@ -1,0 +1,6 @@
+package hwr.oop.group1.poker.cli
+
+interface StateSerializable {
+    fun toState(): Map<String, Any>
+    fun fromState(state: Map<String, Any>)
+}
